renaming components consistently as protobuf definition
